+++ Plagiator.Music/Models/Song.cs
@@ -2,11 +2,11 @@
 {
     public partial class Song
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
-        public int BandId { get; set; }
+        public long BandId { get; set; }
         public Band Band { get; set; }
-        public int StyleId { get; set; }
+        public long StyleId { get; set; }
         public Style Style { get; set; }
 
       //  public int? TicksPerQuarterNote { get; set; }
